refactor(tools): share character normalization helpers and use in-memory DB
